+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The **patch** part changes incrementally at each release.
 
 ### Changed
 
+* nginx: change from "nginx_status-XXX" to "server-status-XXX"
+
 ### Fixed
 
 ### Removed